Move file-operations.ts to file-system/controllers
